Lots of changes to silence warnings and to revert to non-ARC code

+++ Ukelele Cocoa/UKKeyboardPrintView.m
@@ -101,7 +101,7 @@
 }
 
 - (BOOL)knowsPageRange:(NSRangePointer)range {
-	range->length = ceil([[self subviews] count] / (CGFloat)self.printingInfo.viewsPerPage);
+	range->length = (NSUInteger)(ceil([[self subviews] count] / (CGFloat)self.printingInfo.viewsPerPage));
 	return YES;
 }
